can publish sensor info to AWS dynamoDB

+++ src/tempHumSensor.h
@@ -5,8 +5,8 @@
 
 void setupAHT20();
 
-int getTemperature();
+float getTemperature();
 
-int getHumidity();
+float getHumidity();
 
 #endif
